fix compilation in tests
